requirements.txt
Flask==0.12.2
Flask-RESTful==0.3.6
PyGithub==1.35
pyjwt==1.5.3
cryptography==2.1.4
six==1.11.0
